refactor(block-time): proposals, booster, staking, balances (5/7)
