+++ src/app/proposals/hooks/useFetchLatestProposals.ts
@@ -1,20 +1,17 @@
-import { useQuery, UseQueryResult } from '@tanstack/react-query'
-import { useMemo } from 'react'
-import type { Log } from 'viem'
-import { getAddress, parseEventLogs, prepareEncodeFunctionData } from 'viem'
-
-import { fetchProposalsCreatedCached } from '@/app/proposals/actions/fetchProposalsCreatedCached'
-import { ADDRESS_PADDING_LENGTH, RELAY_PARAMETER_PADDING_LENGTH } from '@/app/proposals/shared/utils'
+import { fetchProposalsCreatedCached } from '@/app/user/Balances/actions'
 import { GovernorAbi } from '@/lib/abis/Governor'
 import { SimplifiedRewardDistributorAbi } from '@/lib/abis/SimplifiedRewardDistributorAbi'
+import { useQuery, UseQueryResult } from '@tanstack/react-query'
+import { useMemo } from 'react'
+import { getAddress, parseEventLogs, prepareEncodeFunctionData } from 'viem'
+
+import { ADDRESS_PADDING_LENGTH, RELAY_PARAMETER_PADDING_LENGTH } from '@/app/proposals/shared/utils'
 import { BuilderRegistryAbi } from '@/lib/abis/tok/BuilderRegistryAbi'
-import { AVERAGE_BLOCKTIME } from '@/lib/constants'
 
 const useFetchLatestProposals = () => {
   return useQuery({
     queryFn: fetchProposalsCreatedCached,
     queryKey: ['proposalsCreated'],
-    refetchInterval: AVERAGE_BLOCKTIME,
   })
 }
 
@@ -25,7 +22,7 @@ export const useFetchAllProposals = () => {
     if (data?.data) {
       const proposals = parseEventLogs({
         abi: GovernorAbi,
-        logs: data.data as unknown as Log[],
+        logs: data.data,
         eventName: 'ProposalCreated',
       })
 
@@ -87,7 +84,7 @@ export const useFetchCreateBuilderProposals = (): ProposalQueryResult<ProposalsP
 
     const events = parseEventLogs({
       abi: GovernorAbi,
-      logs: fetchedData.data as unknown as Log[],
+      logs: fetchedData.data,
       eventName: 'ProposalCreated',
     }) as CreateBuilderProposalEventLog[]
