+++ src/app/proposals/hooks/useVoteCast.ts
@@ -1,10 +1,9 @@
-import { useQuery } from '@tanstack/react-query'
-import { useEffect, useState } from 'react'
-import { Address, Log, parseEventLogs, zeroAddress } from 'viem'
-
-import { fetchVoteCastEventByAccountAddress } from '@/app/user/Balances/actions'
 import { governor } from '@/lib/contracts'
+import { useQuery } from '@tanstack/react-query'
+import { Address, Log, parseEventLogs, zeroAddress } from 'viem'
+import { fetchVoteCastEventByAccountAddress } from '@/app/user/Balances/actions'
 import { Vote, VOTES_MAP } from '@/shared/types'
+import { useEffect, useState } from 'react'
 
 export const parseVoteCastEvents = async (address: Address) => {
   try {
@@ -31,12 +30,12 @@ export const useGetVoteForSpecificProposal = (
   address: Address = zeroAddress,
   proposalId: string,
 ): [Vote | undefined, (vote: Vote | undefined) => void] => {
-  const [vote, setVote] = useState<Vote | undefined>()
+  const [vote, setVote] = useState<Vote | undefined>(undefined)
 
   const { data: voteEvents } = useQuery({
     queryFn: () => parseVoteCastEvents(address),
     queryKey: ['VoteCast', address],
-    enabled: !!address && !!proposalId, // Query will only run if both address and proposalId are present
+    enabled: !!address && !!proposalId,
   })
 
   const event = voteEvents?.find(event => event.args.proposalId.toString() === proposalId)
