+++ src/app/proposals/hooks/useVotingPower.ts
@@ -1,16 +1,16 @@
-import { useQuery } from '@tanstack/react-query'
-import { formatUnits } from 'viem'
-import { useAccount, useReadContracts } from 'wagmi'
-
-import { getCachedProposalSharedDetails } from '@/app/proposals/actions/proposals-action'
 import { StRIFTokenAbi } from '@/lib/abis/StRIFTokenAbi'
 import { tokenContracts } from '@/lib/contracts'
+import { formatUnits } from 'viem'
+import { useAccount, useReadContracts } from 'wagmi'
+import { getCachedProposalSharedDetails } from '@/app/proposals/actions/proposalsAction'
+import { useQuery } from '@tanstack/react-query'
 
 export const useVotingPower = () => {
   const { address } = useAccount()
   const { data: proposalDetails, isLoading: isProposalsDetailsLoading } = useQuery({
     queryFn: getCachedProposalSharedDetails,
     queryKey: ['cachedProposalsSharedDetails'],
+    refetchInterval: false,
   })
 
   const { data, isLoading: isLoadingVotes } = useReadContracts({
